rename table_metric.py to table_metric_provider.py with deprecation notice
