+++ great_expectations/expectations/metrics/column_aggregate_metric.py
@@ -21,7 +21,9 @@ from great_expectations.expectations.metrics.metric_provider import (
     metric_partial,
     metric_value,
 )
-from great_expectations.expectations.metrics.table_metric import TableMetricProvider
+from great_expectations.expectations.metrics.table_metric_provider import (
+    TableMetricProvider,
+)
 from great_expectations.validator.validation_graph import MetricConfiguration
 
 logger = logging.getLogger(__name__)
